airflow_dags/airflow_requirements.txt
# Минимальный набор пакетов для Airflow
pandas==2.0.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1
requests==2.31.0
pyarrow==13.0.0
matplotlib==3.8.3
seaborn==0.13.0
mlflow==2.7.1
fastapi==0.104.1
uvicorn==0.24.0.post1
apache-airflow-providers-common-sql==1.8.1
apache-airflow-providers-ftp==3.6.1
apache-airflow-providers-http==4.7.0
apache-airflow-providers-imap==3.4.0
apache-airflow-providers-postgres==5.8.0
apache-airflow-providers-sqlite==3.5.0
apache-airflow-providers-telegram==4.2.0
apache-airflow==2.7.3
catboost==1.2.2
scipy==1.11.3
implicit==0.7.2
scikit-learn==1.7.1
boto3==1.34.78
